add xdai bridge handling
